refactor: move the dagger modules into just injection
Instead of using injection.modules just use injection, since we don't need components anymore
