+++ app/src/androidTest/java/com/chesire/nekome/injection/components/TestComponent.kt
@@ -16,14 +16,14 @@ import com.chesire.nekome.harness.FakeSearchApi
 import com.chesire.nekome.harness.FakeTrendingApi
 import com.chesire.nekome.harness.FakeUrlHandler
 import com.chesire.nekome.harness.FakeUserApi
-import com.chesire.nekome.injection.modules.AccountModule
-import com.chesire.nekome.injection.modules.AppModule
-import com.chesire.nekome.injection.modules.CoroutineModule
+import com.chesire.nekome.injection.AccountModule
+import com.chesire.nekome.injection.AppModule
+import com.chesire.nekome.injection.CoroutineModule
 import com.chesire.nekome.injection.modules.FakeKitsuModule
 import com.chesire.nekome.injection.modules.FakeUrlModule
 import com.chesire.nekome.injection.modules.MemoryDatabaseModule
-import com.chesire.nekome.injection.modules.SeriesModule
-import com.chesire.nekome.injection.modules.ServerModule
+import com.chesire.nekome.injection.SeriesModule
+import com.chesire.nekome.injection.ServerModule
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjector
