+++ app/src/main/java/com/chesire/nekome/injection/AccountModule.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.injection.modules
+package com.chesire.nekome.injection
 
 import com.chesire.nekome.account.UserRepository
 import com.chesire.nekome.database.dao.UserDao
